Refactor Appointment, Add Parser for AddAppointmentCommand
AddAppointmentCommandParser will create a Command containing an Appointment object.
The command will be executed to store into the model and storage.

However, it is not possible to create Client and Service objects within ParserUtil as attributes
for Appointment, as a search is needed to be performed within Model to match the input Phone
and input ServiceCode to the right Client and Service. Thus, let's refactor Appointment to
include Phone and ServiceCode as attributes, while making Client and Service Optional.